refactor : Main에서 처리하던 request요청을 controller로 이동
